+++ src/main/java/de/thexxturboxx/fontfixer/FontFixer.java
@@ -25,7 +25,7 @@ public class FontFixer {
     public Random fontRandom = new Random();
     private byte[] glyphWidth = new byte[65536];
     private final int[] glyphTextureName = new int[256];
-    private int[] colorCode = new int[32];
+    public int[] colorCode = new int[32];
     private int boundTextureName;
     private final TextureManager renderEngine;
     private float posX;
@@ -44,31 +44,31 @@ public class FontFixer {
     private boolean strikethroughStyle = false;
     private final ResourceLocation locationFontTexture;
 
-    private final String ASCII = "ÀÁÂÈÊËÍÓÔÕÚßãõğİıŒœŞşŴŵžȇ\u0000\u0000\u0000\u0000\u0000\u0000\u0000 !\"#$%&'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_`abcdefghijklmnopqrstuvwxyz{|}~\u0000ÇüéâäàåçêëèïîìÄÅÉæÆôöòûùÿÖÜø£Ø×ƒáíóúñÑªº¿®¬½¼¡«»░▒▓│┤╡╢╖╕╣║╗╝╜╛┐└┴┬├─┼╞╟╚╔╩╦╠═╬╧╨╤╥╙╘╒╓╫╪┘┌█▄▌▐▀αβΓπΣσμτΦΘΩδ∞∅∈∩≡±≥≤⌠⌡÷≈°∙·√ⁿ²■\u0000";
+    public final String ASCII = "ÀÁÂÈÊËÍÓÔÕÚßãõğİıŒœŞşŴŵžȇ\u0000\u0000\u0000\u0000\u0000\u0000\u0000 !\"#$%&'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_`abcdefghijklmnopqrstuvwxyz{|}~\u0000ÇüéâäàåçêëèïîìÄÅÉæÆôöòûùÿÖÜø£Ø×ƒáíóúñÑªº¿®¬½¼¡«»░▒▓│┤╡╢╖╕╣║╗╝╜╛┐└┴┬├─┼╞╟╚╔╩╦╠═╬╧╨╤╥╙╘╒╓╫╪┘┌█▄▌▐▀αβΓπΣσμτΦΘΩδ∞∅∈∩≡±≥≤⌠⌡÷≈°∙·√ⁿ²■\u0000";
 
 
 //    public FontFixer() {
 //        this.renderEngine = null;
 //    }
 
-    public FontFixer(GameSettings par1GameSettings, ResourceLocation par2ResourceLocation, TextureManager par3RenderEngine, boolean par4) {
-        this.locationFontTexture = par2ResourceLocation;
-        this.renderEngine = par3RenderEngine;
-        this.unicodeFlag = par4;
-        par3RenderEngine.bindTexture(this.locationFontTexture);
+    public FontFixer(GameSettings gameSettings, ResourceLocation resourceLocation, TextureManager renderEngine, boolean unicodeFlag) {
+        this.locationFontTexture = resourceLocation;
+        this.renderEngine = renderEngine;
+        this.unicodeFlag = unicodeFlag;
+        renderEngine.bindTexture(this.locationFontTexture);
         this.readGlyphSizes();
 
-        BufferedImage var5;
+        BufferedImage bufferedImage;
         try {
-            var5 = ImageIO.read(Minecraft.getMinecraft().getResourceManager().getResource(this.locationFontTexture).getInputStream());
+            bufferedImage = ImageIO.read(Minecraft.getMinecraft().getResourceManager().getResource(this.locationFontTexture).getInputStream());
         } catch (IOException var17) {
             throw new RuntimeException(var17);
         }
 
-        int var19 = var5.getWidth();
-        int var7 = var5.getHeight();
+        int var19 = bufferedImage.getWidth();
+        int var7 = bufferedImage.getHeight();
         int[] var8 = new int[var19 * var7];
-        var5.getRGB(0, 0, var19, var7, var8, 0, var19);
+        bufferedImage.getRGB(0, 0, var19, var7, var8, 0, var19);
 
         int var9;
         int var10;
@@ -106,7 +106,7 @@ public class FontFixer {
         }
 
         this.fontTextureName = 1;
-//                par3RenderEngine.method_1417(var5);
+//        renderEngine.method_1417(bufferedImage);
 
         for (var9 = 0; var9 < 32; ++var9) {
             var10 = (var9 >> 3 & 1) * 85;
@@ -117,7 +117,7 @@ public class FontFixer {
                 var11 += 85;
             }
 
-            if (par1GameSettings.anaglyph) {
+            if (gameSettings.anaglyph) {
                 int var20 = (var11 * 30 + var12 * 59 + var13 * 11) / 100;
                 var15 = (var11 * 30 + var12 * 70) / 100;
                 var16 = (var11 * 30 + var13 * 70) / 100;
@@ -144,7 +144,7 @@ public class FontFixer {
                 var20 += 85;
             }
 
-            if (par1GameSettings.anaglyph) {
+            if (gameSettings.anaglyph) {
                 int var23 = (var20 * 30 + var21 * 59 + var22 * 11) / 100;
                 int var24 = (var20 * 30 + var21 * 70) / 100;
                 int var25 = (var20 * 30 + var22 * 70) / 100;
@@ -165,29 +165,29 @@ public class FontFixer {
 
     private void readGlyphSizes() {
         try {
-            InputStream inputstream = getResourceInputStream(new ResourceLocation("font/glyph_sizes.bin"));
-            inputstream.read(this.glyphWidth);
+            InputStream stream = Minecraft.getMinecraft().getResourceManager().getResource(new ResourceLocation("font/glyph_sizes.bin")).getInputStream();
+            stream.read(this.glyphWidth);
             this.glyphWidth['（'] = 127;
         } catch (IOException var2) {
             throw new RuntimeException(var2);
         }
     }
 
-    private float renderCharAtPos(int par1, char par2, boolean par3) {
-        if (par2 == ' ') {
+    public float renderCharAtPos(int charWidth, char glyphWidth, boolean par3) {
+        if (glyphWidth == ' ') {
             return 4.0F;
         } else {
-            return (this.ASCII.indexOf(par2) != -1) && !BGSConfig.FORCE_UNICODE_FONT.getValue() ? this.renderDefaultChar(par1, par3) : this.renderUnicodeChar(par2, par3);
+            return (this.ASCII.indexOf(glyphWidth) != -1) && !BGSConfig.FORCE_UNICODE_FONT.getValue() ? this.renderDefaultChar(charWidth, par3) : this.renderUnicodeChar(glyphWidth, par3);
         }
     }
 
     private float renderDefaultChar(int par1, boolean par2) {
-        float var3 = (float) (par1 % 16 * 8);
-        float var4 = (float) (par1 / 16 * 8);
+        float var3 = (float)(par1 % 16 * 8);
+        float var4 = (float)(par1 / 16 * 8);
         float var5 = par2 ? 1.0F : 0.0F;
         this.renderEngine.bindTexture(this.locationFontTexture);
 
-        float var6 = (float) this.charWidth[par1] - 0.01F;
+        float var6 = (float)this.charWidth[par1] - 0.01F;
         GL11.glBegin(5);
         GL11.glTexCoord2f(var3 / 128.0F, var4 / 128.0F);
         GL11.glVertex3f(this.posX + var5, this.posY, 0.0F);
@@ -198,7 +198,7 @@ public class FontFixer {
         GL11.glTexCoord2f((var3 + var6) / 128.0F, (var4 + 7.99F) / 128.0F);
         GL11.glVertex3f(this.posX + var6 - var5, this.posY + 7.99F, 0.0F);
         GL11.glEnd();
-        return (float) this.charWidth[par1];
+        return (float)this.charWidth[par1];
     }
 
     private ResourceLocation getUnicodePageLocation(int page) {
@@ -209,22 +209,22 @@ public class FontFixer {
         return unicodePageLocations[page];
     }
 
-    public final void loadGlyphTexture(int par1) {
-        this.renderEngine.bindTexture(this.getUnicodePageLocation(par1));
+    public final void loadGlyphTexture(int page) {
+        this.renderEngine.bindTexture(this.getUnicodePageLocation(page));
     }
 
-    private float renderUnicodeChar(char par1, boolean par2) {
-        if (this.glyphWidth[par1] == 0) {
+    private float renderUnicodeChar(char glyphWidth, boolean par2) {
+        if (this.glyphWidth[glyphWidth] == 0) {
             return 0.0F;
         } else {
-            int var3 = par1 / 256;
+            int var3 = glyphWidth / 256;
             this.loadGlyphTexture(var3);
-            int var4 = this.glyphWidth[par1] >>> 4;
-            int var5 = this.glyphWidth[par1] & 15;
-            float var6 = (float) var4;
-            float var7 = (float) (var5 + 1);
-            float var8 = (float) (par1 % 16 * 16) + var6;
-            float var9 = (float) ((par1 & 255) / 16 * 16);
+            int var4 = this.glyphWidth[glyphWidth] >>> 4;
+            int var5 = this.glyphWidth[glyphWidth] & 15;
+            float var6 = (float)var4;
+            float var7 = (float)(var5 + 1);
+            float var8 = (float)(glyphWidth % 16 * 16) + var6;
+            float var9 = (float)((glyphWidth & 255) / 16 * 16);
             float var10 = var7 - var6 - 0.02F;
             float var11 = par2 ? 1.0F : 0.0F;
             GL11.glBegin(5);
@@ -241,57 +241,57 @@ public class FontFixer {
         }
     }
 
-    public int drawStringWithShadow(String par1Str, int par2, int par3, int par4) {
-        return this.drawString(par1Str, par2, par3, par4, true);
+    public int drawStringWithShadow(String text, int x, int y, int textColor) {
+        return this.drawString(text, x, y, textColor, true);
     }
 
-    public int drawString(String par1Str, int par2, int par3, int par4) {
+    public int drawString(String text, int x, int y, int textColor) {
+        return this.drawString(text, x, y, textColor, false);
+    }
+
+    public int drawString(String text, int x, int y, int textColor, boolean shadow) {
         GL11.glEnable(GL11.GL_ALPHA_TEST);
-        return this.drawString(par1Str, par2, par3, par4, false);
-    }
-
-    public int drawString(String par1Str, int par2, int par3, int par4, boolean par5) {
         this.resetStyles();
         if (this.bidiFlag) {
-            par1Str = this.bidiReorder(par1Str);
+            text = this.bidiReorder(text);
         }
 
         int var6;
-        if (par5) {
-            var6 = this.renderString(par1Str, par2 + 1, par3 + 1, par4, true);
-            var6 = Math.max(var6, this.renderString(par1Str, par2, par3, par4, false));
+        if (shadow) {
+            var6 = this.renderString(text, x + 1, y + 1, textColor, true);
+            var6 = Math.max(var6, this.renderString(text, x, y, textColor, false));
         } else {
-            var6 = this.renderString(par1Str, par2, par3, par4, false);
+            var6 = this.renderString(text, x, y, textColor, false);
         }
 
         return var6;
     }
 
-    private String bidiReorder(String par1Str) {
-        if (par1Str != null && Bidi.requiresBidi(par1Str.toCharArray(), 0, par1Str.length())) {
-            Bidi var2 = new Bidi(par1Str, -2);
+    private String bidiReorder(String text) {
+        if (text != null && Bidi.requiresBidi(text.toCharArray(), 0, text.length())) {
+            Bidi var2 = new Bidi(text, -2);
             byte[] var3 = new byte[var2.getRunCount()];
             String[] var4 = new String[var3.length];
 
             int var7;
-            for (int var5 = 0; var5 < var3.length; ++var5) {
+            for(int var5 = 0; var5 < var3.length; ++var5) {
                 int var6 = var2.getRunStart(var5);
                 var7 = var2.getRunLimit(var5);
                 int var8 = var2.getRunLevel(var5);
-                String var9 = par1Str.substring(var6, var7);
-                var3[var5] = (byte) var8;
+                String var9 = text.substring(var6, var7);
+                var3[var5] = (byte)var8;
                 var4[var5] = var9;
             }
 
-            String[] var11 = (String[]) var4.clone();
+            String[] var11 = (String[])var4.clone();
             Bidi.reorderVisually(var3, 0, var4, 0, var3.length);
             StringBuilder var12 = new StringBuilder();
 
-            for (var7 = 0; var7 < var4.length; ++var7) {
+            for(var7 = 0; var7 < var4.length; ++var7) {
                 byte var13 = var3[var7];
 
                 int var14;
-                for (var14 = 0; var14 < var11.length; ++var14) {
+                for(var14 = 0; var14 < var11.length; ++var14) {
                     if (var11[var14].equals(var4[var7])) {
                         var13 = var3[var14];
                         break;
@@ -301,7 +301,7 @@ public class FontFixer {
                 if ((var13 & 1) == 0) {
                     var12.append(var4[var7]);
                 } else {
-                    for (var14 = var4[var7].length() - 1; var14 >= 0; --var14) {
+                    for(var14 = var4[var7].length() - 1; var14 >= 0; --var14) {
                         char var10 = var4[var7].charAt(var14);
                         if (var10 == '(') {
                             var10 = ')';
@@ -316,7 +316,7 @@ public class FontFixer {
 
             return var12.toString();
         } else {
-            return par1Str;
+            return text;
         }
     }
 
@@ -328,13 +328,13 @@ public class FontFixer {
         this.strikethroughStyle = false;
     }
 
-    private void renderStringAtPos(String string, boolean shadow) {
-        for (int var3 = 0; var3 < string.length(); ++var3) {
-            char var4 = string.charAt(var3);
+    public void renderStringAtPos(String text, boolean shadow) {
+        for(int var3 = 0; var3 < text.length(); ++var3) {
+            char var4 = text.charAt(var3);
             int var5;
             int var6;
-            if (var4 == 167 && var3 + 1 < string.length()) {
-                var5 = "0123456789abcdefklmnor".indexOf(string.toLowerCase().charAt(var3 + 1));
+            if (var4 == 167 && var3 + 1 < text.length()) {
+                var5 = "0123456789abcdefklmnor".indexOf(text.toLowerCase().charAt(var3 + 1));
                 if (var5 < 16) {
                     this.randomStyle = false;
                     this.boldStyle = false;
@@ -351,8 +351,8 @@ public class FontFixer {
 
                     var6 = this.colorCode[var5];
                     this.textColor = var6;
-                    GL11.glColor4f((float) (var6 >> 16) / 255.0F, (float) (var6 >> 8 & 255) / 255.0F, (float) (var6 & 255) / 255.0F, this.alpha);
-                    var3 = applyCustomFormatCodes(this, string, shadow, var3);
+                    GL11.glColor4f((float)(var6 >> 16) / 255.0F, (float)(var6 >> 8 & 255) / 255.0F, (float)(var6 & 255) / 255.0F, this.alpha);
+                    var3 = applyCustomFormatCodes(this, text, shadow, var3);
                 } else if (var5 == 16) {
                     this.randomStyle = true;
                 } else if (var5 == 17) {
@@ -418,10 +418,10 @@ public class FontFixer {
                     var7 = Tessellator.instance;
                     GL11.glDisable(3553);
                     var7.startDrawingQuads();
-                    var7.addVertex((double) this.posX, (double) (this.posY + (float) (this.FONT_HEIGHT / 2)), 0.0);
-                    var7.addVertex((double) (this.posX + var9), (double) (this.posY + (float) (this.FONT_HEIGHT / 2)), 0.0);
-                    var7.addVertex((double) (this.posX + var9), (double) (this.posY + (float) (this.FONT_HEIGHT / 2) - 1.0F), 0.0);
-                    var7.addVertex((double) this.posX, (double) (this.posY + (float) (this.FONT_HEIGHT / 2) - 1.0F), 0.0);
+                    var7.addVertex((double)this.posX, (double)(this.posY + (float)(this.FONT_HEIGHT / 2)), 0.0);
+                    var7.addVertex((double)(this.posX + var9), (double)(this.posY + (float)(this.FONT_HEIGHT / 2)), 0.0);
+                    var7.addVertex((double)(this.posX + var9), (double)(this.posY + (float)(this.FONT_HEIGHT / 2) - 1.0F), 0.0);
+                    var7.addVertex((double)this.posX, (double)(this.posY + (float)(this.FONT_HEIGHT / 2) - 1.0F), 0.0);
                     var7.draw();
                     GL11.glEnable(3553);
                 }
@@ -431,10 +431,10 @@ public class FontFixer {
                     GL11.glDisable(3553);
                     var7.startDrawingQuads();
                     int var8 = this.underlineStyle ? -1 : 0;
-                    var7.addVertex((double) (this.posX + (float) var8), (double) (this.posY + (float) this.FONT_HEIGHT), 0.0);
-                    var7.addVertex((double) (this.posX + var9), (double) (this.posY + (float) this.FONT_HEIGHT), 0.0);
-                    var7.addVertex((double) (this.posX + var9), (double) (this.posY + (float) this.FONT_HEIGHT - 1.0F), 0.0);
-                    var7.addVertex((double) (this.posX + (float) var8), (double) (this.posY + (float) this.FONT_HEIGHT - 1.0F), 0.0);
+                    var7.addVertex((double)(this.posX + (float)var8), (double)(this.posY + (float)this.FONT_HEIGHT), 0.0);
+                    var7.addVertex((double)(this.posX + var9), (double)(this.posY + (float)this.FONT_HEIGHT), 0.0);
+                    var7.addVertex((double)(this.posX + var9), (double)(this.posY + (float)this.FONT_HEIGHT - 1.0F), 0.0);
+                    var7.addVertex((double)(this.posX + (float)var8), (double)(this.posY + (float)this.FONT_HEIGHT - 1.0F), 0.0);
                     var7.draw();
                     GL11.glEnable(3553);
                 }
@@ -448,11 +448,11 @@ public class FontFixer {
     /**
      * api for emi
      */
-    public static int applyCustomFormatCodes(FontFixer subject, String str, boolean shadow, int i) {
-        if (str.charAt(i + 1) == 'x') {
-            int next = str.indexOf(String.valueOf('\u00a7') + "x", i + 1);
+    public static int applyCustomFormatCodes(FontFixer subject, String text, boolean shadow, int i) {
+        if (text.charAt(i + 1) == 'x') {
+            int next = text.indexOf(String.valueOf('\u00a7') + "x", i + 1);
             if (next != -1) {
-                String s = str.substring(i + 1, next);
+                String s = text.substring(i + 1, next);
                 int color = Integer.parseInt(s.replace(String.valueOf('\u00a7'), "").substring(1), 16);
                 if (shadow) {
                     color = (color & 16579836) >> 2 | color & -16777216;
@@ -465,57 +465,57 @@ public class FontFixer {
         return i;
     }
 
-    private int renderStringAligned(String par1Str, int par2, int par3, int par4, int par5, boolean par6) {
+    private int renderStringAligned(String text, int x, int y, int width, int textColor, boolean shadow) {
         if (this.bidiFlag) {
-            par1Str = this.bidiReorder(par1Str);
-            int var7 = this.getStringWidth(par1Str);
-            par2 = par2 + par4 - var7;
+            text = this.bidiReorder(text);
+            int var7 = this.getStringWidth(text);
+            x = x + width - var7;
         }
 
-        return this.renderString(par1Str, par2, par3, par5, par6);
+        return this.renderString(text, x, y, textColor, shadow);
     }
 
-    private int renderString(String par1Str, int par2, int par3, int par4, boolean par5) {
-        if (par1Str != null && this.bidiFlag) {
-            par1Str = this.bidiReorder(par1Str);
+    private int renderString(String text, int x, int y, int textColor, boolean shadow) {
+        if (text != null && this.bidiFlag) {
+            text = this.bidiReorder(text);
         }
-        if (par1Str == null) {
+        if (text == null) {
             return 0;
         } else {
             this.boundTextureName = 0;
-            if ((par4 & -67108864) == 0) {
-                par4 |= -16777216;
+            if ((textColor & 0xFC000000) == 0) {
+                textColor |= 0xFF000000;
             }
 
-            if (par5) {
-                par4 = (par4 & 16579836) >> 2 | par4 & -16777216;
+            if (shadow) {
+                textColor = (textColor & 0xFCFCFC) >> 2 | textColor & 0xFF000000;
             }
 
-            this.red = (float) (par4 >> 16 & 255) / 255.0F;
-            this.blue = (float) (par4 >> 8 & 255) / 255.0F;
-            this.green = (float) (par4 & 255) / 255.0F;
-            this.alpha = (float) (par4 >> 24 & 255) / 255.0F;
+            this.red = (float) (textColor >> 16 & 0xFF) / 255.0F;
+            this.blue = (float) (textColor >> 8 & 0xFF) / 255.0F;
+            this.green = (float) (textColor & 0xFF) / 255.0F;
+            this.alpha = (float) (textColor >> 24 & 0xFF) / 255.0F;
             GL11.glColor4f(this.red, this.blue, this.green, this.alpha);
-            this.posX = (float) par2;
-            this.posY = (float) par3;
-            this.renderStringAtPos(par1Str, par5);
+            this.posX = (float) x;
+            this.posY = (float) y;
+            this.renderStringAtPos(text, shadow);
             return (int) this.posX;
         }
     }
 
-    public int getStringWidth(String par1Str) {
-        if (par1Str == null) {
+    public int getStringWidth(String text) {
+        if (text == null) {
             return 0;
         } else {
             int var2 = 0;
             boolean var3 = false;
 
-            for (int var4 = 0; var4 < par1Str.length(); ++var4) {
-                char var5 = par1Str.charAt(var4);
+            for (int var4 = 0; var4 < text.length(); ++var4) {
+                char var5 = text.charAt(var4);
                 int var6 = this.getCharWidth(var5);
-                if (var6 < 0 && var4 < par1Str.length() - 1) {
+                if (var6 < 0 && var4 < text.length() - 1) {
                     ++var4;
-                    var5 = par1Str.charAt(var4);
+                    var5 = text.charAt(var4);
                     if (var5 != 'l' && var5 != 'L') {
                         if (var5 == 'r' || var5 == 'R') {
                             var3 = false;
@@ -537,18 +537,18 @@ public class FontFixer {
         }
     }
 
-    public int getCharWidth(char par1) {
-        if (par1 == 167) {
+    public int getCharWidth(char glyphWidth) {
+        if (glyphWidth == 167) {
             return -1;
-        } else if (par1 == ' ') {
+        } else if (glyphWidth == ' ') {
             return 4;
         } else {
-            int var2 = this.ASCII.indexOf(par1);
-            if (par1 > 0 && var2 != -1 && !BGSConfig.FORCE_UNICODE_FONT.getValue()) {
+            int var2 = this.ASCII.indexOf(glyphWidth);
+            if (glyphWidth > 0 && var2 != -1 && !BGSConfig.FORCE_UNICODE_FONT.getValue()) {
                 return this.charWidth[var2];
-            } else if (this.glyphWidth[par1] != 0) {
-                int var3 = this.glyphWidth[par1] >>> 4;
-                int var4 = this.glyphWidth[par1] & 15;
+            } else if (this.glyphWidth[glyphWidth] != 0) {
+                int var3 = this.glyphWidth[glyphWidth] >>> 4;
+                int var4 = this.glyphWidth[glyphWidth] & 15;
                 if (var4 > 7) {
                     var4 = 15;
                     var3 = 0;
@@ -557,25 +557,25 @@ public class FontFixer {
                 ++var4;
                 return (var4 - var3) / 2 + 1;
             } else {
-                return 4;
+                return 0;
             }
         }
     }
 
-    public String trimStringToWidth(String par1Str, int par2) {
-        return this.trimStringToWidth(par1Str, par2, false);
+    public String trimStringToWidth(String text, int width) {
+        return this.trimStringToWidth(text, width, false);
     }
 
-    public String trimStringToWidth(String par1Str, int par2, boolean par3) {
+    public String trimStringToWidth(String text, int width, boolean par3) {
         StringBuilder var4 = new StringBuilder();
         int var5 = 0;
-        int var6 = par3 ? par1Str.length() - 1 : 0;
+        int var6 = par3 ? text.length() - 1 : 0;
         int var7 = par3 ? -1 : 1;
         boolean var8 = false;
         boolean var9 = false;
 
-        for (int var10 = var6; var10 >= 0 && var10 < par1Str.length() && var5 < par2; var10 += var7) {
-            char var11 = par1Str.charAt(var10);
+        for (int var10 = var6; var10 >= 0 && var10 < text.length() && var5 < width; var10 += var7) {
+            char var11 = text.charAt(var10);
             int var12 = this.getCharWidth(var11);
             if (var8) {
                 var8 = false;
@@ -595,7 +595,7 @@ public class FontFixer {
                 }
             }
 
-            if (var5 > par2) {
+            if (var5 > width) {
                 break;
             }
 
@@ -609,72 +609,76 @@ public class FontFixer {
         return var4.toString();
     }
 
-    private String trimStringNewline(String par1Str) {
-        while (par1Str != null && par1Str.endsWith("\n")) {
-            par1Str = par1Str.substring(0, par1Str.length() - 1);
+    private String trimStringNewline(String text) {
+        while(text != null && text.endsWith("\n")) {
+            text = text.substring(0, text.length() - 1);
         }
 
-        return par1Str;
+        return text;
     }
 
-    public void drawSplitString(String par1Str, int par2, int par3, int par4, int par5) {
+    public void drawSplitString(String text, int x, int y, int width, int textColor) {
         this.resetStyles();
-        this.textColor = par5;
-        par1Str = this.trimStringNewline(par1Str);
-        this.renderSplitString(par1Str, par2, par3, par4, false);
+        this.textColor = textColor;
+        text = this.trimStringNewline(text);
+        this.renderSplitString(text, x, y, width, false);
     }
 
-    private void renderSplitString(String par1Str, int par2, int par3, int par4, boolean par5) {
-        List var6 = this.listFormattedStringToWidth(par1Str, par4);
+    private void renderSplitString(String text, int x, int y, int width, boolean shadow) {
+        List textList = this.listFormattedStringToWidth(text, width);
 
-        for (Iterator var7 = var6.iterator(); var7.hasNext(); par3 += this.FONT_HEIGHT) {
+        for (Iterator var7 = textList.iterator(); var7.hasNext(); y += this.FONT_HEIGHT) {
             String var8 = (String) var7.next();
-            this.renderStringAligned(var8, par2, par3, par4, this.textColor, par5);
+            this.renderStringAligned(var8, x, y, width, this.textColor, shadow);
         }
 
     }
 
-    public int splitStringWidth(String par1Str, int par2) {
-        return this.FONT_HEIGHT * this.listFormattedStringToWidth(par1Str, par2).size();
+    public int splitStringWidth(String text, int width) {
+        return this.FONT_HEIGHT * this.listFormattedStringToWidth(text, width).size();
     }
 
-    public void setUnicodeFlag(boolean par1) {
-        this.unicodeFlag = par1;
+    public void setUnicodeFlag(boolean unicodeFlag) {
+        this.unicodeFlag = unicodeFlag;
     }
 
     public boolean getUnicodeFlag() {
         return this.unicodeFlag;
     }
 
-    public void setBidiFlag(boolean par1) {
-        this.bidiFlag = par1;
+    public void setBidiFlag(boolean bidiFlag) {
+        this.bidiFlag = bidiFlag;
     }
 
-    public List listFormattedStringToWidth(String par1Str, int par2) {
-        return Arrays.asList(this.wrapFormattedStringToWidth(par1Str, par2).split("\n"));
+    public boolean getBidiFlag() {
+        return this.bidiFlag;
     }
 
-    public String wrapFormattedStringToWidth(String par1Str, int par2) {
-        int var3 = this.sizeStringToWidth(par1Str, par2);
-        if (par1Str.length() <= var3) {
-            return par1Str;
+    public List listFormattedStringToWidth(String text, int width) {
+        return Arrays.asList(this.wrapFormattedStringToWidth(text, width).split("\n"));
+    }
+
+    public String wrapFormattedStringToWidth(String text, int width) {
+        int var3 = this.sizeStringToWidth(text, width);
+        if (text.length() <= var3) {
+            return text;
         } else {
-            String var4 = par1Str.substring(0, var3);
-            char var5 = par1Str.charAt(var3);
+            String var4 = text.substring(0, var3);
+            char var5 = text.charAt(var3);
             boolean var6 = var5 == ' ' || var5 == '\n';
-            String var7 = getFormatFromString(var4) + par1Str.substring(var3 + (var6 ? 1 : 0));
-            return var4 + "\n" + this.wrapFormattedStringToWidth(var7, par2);
+            String var7 = getFormatFromString(var4) + text.substring(var3 + (var6 ? 1 : 0));
+            return var4 + "\n" + this.wrapFormattedStringToWidth(var7, width);
         }
     }
 
-    private int sizeStringToWidth(String par1Str, int par2) {
-        int var3 = par1Str.length();
+    private int sizeStringToWidth(String text, int width) {
+        int var3 = text.length();
         int var4 = 0;
         int var5 = 0;
         int var6 = -1;
 
-        for (boolean var7 = false; var5 < var3; ++var5) {
-            char var8 = par1Str.charAt(var5);
+        for(boolean var7 = false; var5 < var3; ++var5) {
+            char var8 = text.charAt(var5);
             switch (var8) {
                 case '\n':
                     --var5;
@@ -690,7 +694,7 @@ public class FontFixer {
                 case '§':
                     if (var5 < var3 - 1) {
                         ++var5;
-                        char var9 = par1Str.charAt(var5);
+                        char var9 = text.charAt(var5);
                         if (var9 != 'l' && var9 != 'L') {
                             if (var9 == 'r' || var9 == 'R' || isFormatColor(var9)) {
                                 var7 = false;
@@ -707,7 +711,7 @@ public class FontFixer {
                 break;
             }
 
-            if (var4 > par2) {
+            if (var4 > width) {
                 break;
             }
         }
@@ -715,22 +719,22 @@ public class FontFixer {
         return var5 != var3 && var6 != -1 && var6 < var5 ? var6 : var5;
     }
 
-    private static boolean isFormatColor(char par0) {
-        return par0 >= '0' && par0 <= '9' || par0 >= 'a' && par0 <= 'f' || par0 >= 'A' && par0 <= 'F';
+    private static boolean isFormatColor(char textChar) {
+        return textChar >= '0' && textChar <= '9' || textChar >= 'a' && textChar <= 'f' || textChar >= 'A' && textChar <= 'F';
     }
 
-    private static boolean isFormatSpecial(char par0) {
-        return par0 >= 'k' && par0 <= 'o' || par0 >= 'K' && par0 <= 'O' || par0 == 'r' || par0 == 'R';
+    private static boolean isFormatSpecial(char textChar) {
+        return textChar >= 'k' && textChar <= 'o' || textChar >= 'K' && textChar <= 'O' || textChar == 'r' || textChar == 'R';
     }
 
-    private static String getFormatFromString(String par0Str) {
+    private static String getFormatFromString(String text) {
         StringBuilder var1 = new StringBuilder();
         int var2 = -1;
-        int var3 = par0Str.length();
+        int var3 = text.length();
 
-        while ((var2 = par0Str.indexOf(167, var2 + 1)) != -1) {
+        while((var2 = text.indexOf(167, var2 + 1)) != -1) {
             if (var2 < var3 - 1) {
-                char var4 = par0Str.charAt(var2 + 1);
+                char var4 = text.charAt(var2 + 1);
                 if (isFormatColor(var4)) {
                     var1 = new StringBuilder("§" + var4);
                 } else if (isFormatSpecial(var4)) {
@@ -740,13 +744,5 @@ public class FontFixer {
         }
 
         return var1.toString();
-    }
-
-    public boolean getBidiFlag() {
-        return this.bidiFlag;
-    }
-
-    protected InputStream getResourceInputStream(ResourceLocation location) throws IOException {
-        return Minecraft.getMinecraft().getResourceManager().getResource(location).getInputStream();
     }
 }