+++ ChangeLog.md
@@ -1,5 +1,14 @@
 # Better Game Setting for BTW
 
+### v1.2.1
+* Added input method fix mod( https://github.com/lss233/InputMethodBlocker )
+* Optimized the save of game settings information in this mod, and now no longer overwrites vanilla entries
+* Force turn off advanced OpenGL when using a non-Nvidia GPU
+* Fixed an issue where FontFixer could not be enabled properly
+* Fixed an issue where the opening parenthesis "(" was not displayed
+
+---
+
 ### v1.2.0
 * Added video settings in 1.7.2+
     * For compatibility (Shaders Mod Core), MIPMAP and various heterosexual filtering options are not added